refactor: modelos del CRUD de configuracion, revisados/agrupados
