--- src/app/@core/data/models/aplicacion.ts
@@ -1,10 +1,0 @@
-
-export class Aplicacion {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-  Dominio: string;
-  Estado: boolean;
-  Alias: string;
-  EstiloIcono: string;
-}
