--- src/app/@core/data/models/menu_opcion.ts
@@ -1,16 +1,0 @@
-import { Aplicacion } from './aplicacion';
-
-export class OpcionTipoOpcion {
-  Id: number;
-  Nombre: string;
-}
-
-export class MenuOpcion {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-  Url: string;
-  Layout: string;
-  Aplicacion: Array<Aplicacion>;
-  TipoOpcion: OpcionTipoOpcion;
-}
