--- src/app/@core/data/models/menu_opcion_padre.ts
@@ -1,7 +1,0 @@
-import { MenuOpcion } from './menu_opcion';
-
-export class MenuOpcionPadre {
-  Id: number;
-  Padre: Array<MenuOpcion>;
-  Hijo: Array<MenuOpcion>;
-}
