--- src/app/@core/data/models/metodo_http.ts
@@ -1,6 +1,0 @@
-
-export class MetodoHttp {
-  Id: number;
-  Nombre: string;
-  Descripcion: string;
-}
