--- src/app/@core/data/models/notificacion.ts
@@ -1,8 +1,0 @@
-import { NotificacionConfiguracion } from './notificacion_configuracion';
-
-export class Notificacion {
-  Id: number;
-  FechaCreacion: Date;
-  CuerpoNotificacion: string;
-  NotificacionConfiguracion: NotificacionConfiguracion;
-}
