--- src/app/@core/data/models/notificacion_configuracion.ts
@@ -1,12 +1,0 @@
-import { MetodoHttp } from './metodo_http';
-import { NotificacionTipo } from './notificacion_tipo';
-import { Aplicacion } from './aplicacion';
-
-export class NotificacionConfiguracion {
-  Id: number;
-  EndPoint: string;
-  MetodoHttp: MetodoHttp;
-  Tipo: NotificacionTipo;
-  CuerpoNotificacion: string;
-  Aplicacion: Aplicacion;
-}
