--- src/app/@core/data/models/notificacion_configuracion_perfil.ts
@@ -1,8 +1,0 @@
-import { NotificacionConfiguracion } from './notificacion_configuracion';
-import { Perfil } from './perfil';
-
-export class NotificacionConfiguracionPerfil {
-  Id: number;
-  NotificacionConfiguracion: NotificacionConfiguracion;
-  Perfil: Perfil;
-}
