--- src/app/@core/data/models/notificacion_estado.ts
@@ -1,9 +1,0 @@
-
-export class NotificacionEstado {
-  Id: number;
-  Nombre: string;
-  CodigoAbreviacion: string;
-  Descripcion: string;
-  Activo: boolean;
-  NumeroOrden: string;
-}
