--- src/app/@core/data/models/notificacion_estado_usuario.ts
@@ -1,11 +1,0 @@
-import { Notificacion } from './notificacion';
-import { NotificacionEstado } from './notificacion_estado';
-
-export class NotificacionEstadoUsuario {
-  Id: number;
-  Notificacion: Array<Notificacion>;
-  NotificacionEstado: Array<NotificacionEstado>;
-  Fecha: Date;
-  Usuario: string;
-  Activo: boolean;
-}
