--- src/app/@core/data/models/notificacion_tipo.ts
@@ -1,5 +1,0 @@
-
-export class NotificacionTipo {
-  Id: number;
-  Nombre: string;
-}
