--- src/app/@core/data/models/parametro.ts
@@ -1,8 +1,0 @@
-import { Aplicacion } from './aplicacion';
-
-export class Parametro {
-  Id: number;
-  Nombre: string;
-  Valor: string;
-  Aplicacion: Array<Aplicacion>;
-}
