--- src/app/@core/data/models/perfil.ts
@@ -1,7 +1,0 @@
-import { Aplicacion } from './aplicacion';
-
-export class Perfil {
-  Id: number;
-  Nombre: string;
-  Aplicacion: Aplicacion;
-}
