--- src/app/@core/data/models/perfil_x_menu_opcion.ts
@@ -1,7 +1,0 @@
-import { MenuOpcion } from './menu_opcion';
-
-export class PerfilXMenuOpcion {
-  Id: number;
-  Perfil: number;
-  Opcion: Array<MenuOpcion>;
-}
